requirements.txt
ChatGPT_lite==0.0.5
gtts==2.3.1
python-dotenv==0.21.1
SpeechRecognition==3.8.1
wikipedia==1.4.0
